Show error when Dart SDK is not set correctly (#646)
Fixes #636

Since the Flutter SDK is derived the from Dart SDK, we actually handle
this by reporting: 'No Flutter SDK configured'. When the user fixes
the Flutter SDK, the Dart SDK will also be fixed.

* Fixed combo box to display correctly when no Flutter SDK is configured.
* Fire a Flutter SDK changed event when the user changes the Flutter SDK.
* Removed dead code.
* fix race condition with Flutter version in dialog

Not done: when the user changes the Dart SDK, we should fire a Flutter
SDK changed event.