Add support for loading a plugin config file from a Bazel Workspace
Finds the workspace based on the content roots. The ancestor directories of
each content root are searched for a WORKSPACE file. There should be a unique
workspace for each project.

Also, automatically detect changes to the configuration. This can be due
to a changed WORKSPACE file, config file, or a changed module configuration.

Also, implemented some test infrastructure to test it all, based on JUnit 4.

Not hooked up to anything yet.