device list visibility fixes for Bazel (#776)
For Bazel, only show the device list when we are in a
Flutter project (according to the Workspace).

If the Bazel repo has a script to start the device
daemon, prefer that, but fall back to the Flutter SDK
if configured.

Fixes a bug where the device menu would not show up
because the user configured a Flutter SDK (for open
source work) but forgot they did that. The device list
should still appear.

Also, log when the device daemon starts and stops
(at INFO level).