Add new run configuration type for 'flutter test' (#1069)
For now this just executes 'flutter test' and sends the output
to the console. Still to do: show test output in IDEA's UI
and support debugging.

Previously, running a test from within IDEA would fail because
the Dart plugin does it and it doesn't use "flutter test",
which is needed for all tests using a flutter pubspec.