Refactor of how metrics are calculated to make adding additional metrics easier.
Refactor of FlutterWidgetPerf so that widget names are always available instead
of only available when displaying gutter icons. The problem was we were calculating widget names for given offsets
when generating the EditorPerfDecorations instead of keeping a global cache
of widget names. The new approach is more efficient and needed for
work to display a table of the most frequent widget rebuilds.