Make JX Browser the fallback and remove settings option (#7232)
This removes the option to select JCEF or JX Browser in settings for
simplicity. We will instead check if JCEF is available first, and if
not, download and use JX Browser. Embedded browser failures in IJ (where
JCEF is included by default) were very low, so I think it's unlikely
users will benefit from having a choice here.

I still need to build this to check whether this is detecting the
presence of JCEF accurately.