Delete unused inspector code from FlutterView (#7414)
Some of the remaining classes that I was going to update with action
thread BGT are related to the legacy inspector that's no longer in use.
I was going to delete more code, but some inspector panel code and
actions seem to be tied up with the performance panel, which is still
available. So I'm going to delete this code but leave the rest of
inspector to delete later, and probably skip changing action threads
since those classes seem inaccessible for testing anyway.