Add a separate file for plugin logs (#8253)
This adds a new class `PluginLogger` that prints log messages to a
separate `flutter.log` file. Log messages will continue to print to
`idea.log` also because `PluginLogger` has a parent logger that
continues to log there.

The built in `FileHandler` class manages writing to the file and
removing old lines if the file gets too big.