Clean up artifact dependencies, jx browser dependency (#8326)
The change to remove dependencies on files in the `artifacts` directory
could break something I'm not aware of, but the plugin seems to build
successfully (both from `bin/plugin make` and on sandbox) and
`bin/plugin test` also runs successfully. My suspicion is that the
artifacts directory used to exist but no longer exists.

The 3 previous JX Browser dependencies seem to be equivalent to
`implementation` so I changed to that.
([doc](https://docs.gradle.org/current/userguide/dependency_configurations.html#sub:what-are-dependency-configurations))