Rename "observatoryUri" to "vmServiceUri" for test.startedProcess (#8328)
This is due to the removal of Observatory support from Dart SDKs and
Flutter which made the tests never unpause on recent versions of
Flutter. This fixes #8233.

The URI pointing to Observatory is no longer given in
`test.startedProcess`. Now, the URI to the VM Service for DevTools is
present and can be used as a drop-in replacement to achieve the same
functionality.