Change the message sent to IntelliJ when starting the test suites to the correct one (#8330)
This is the cause of the error mentioned in #8325.
Previously, the correct message was being sent hard-coded and was
supposed to be replaced by `ServiceMessageBuilder.testsStarted()` once
the method was available. Instead,
`ServiceMessageBuilder.testStarted("")` was used which changes the
message to `##teamcity[testStarted name='']` from
`##teamcity[enteredtheMatrix]` (hence the error).

This one-liner just replaces the erroneous call to the correct one.