Change editor directory files to use PluginLogger (#8372)
Related to https://github.com/flutter/flutter-intellij/issues/8369

I also attempted to make the flutter.log file viewable in sandbox mode
with a generic file path, but this doesn't work for now - I get an empty
file. Oddly this file path will work for `idea.log`, but seems to skip
some lines. As the note says, this topic needs further investigation to
understand how to optimize this part of setup.