Move log level setup to initializer (#8564)
Fixes https://github.com/flutter/flutter-intellij/issues/8562

I'm realizing from this problem that logging at plugin startup may be
tricky. E.g. with this change, I'm not totally sure what the setting
will be at project initialization - probably with a level of INFO, or
whatever is default in the IntelliJ logger. That's probably fine - if we
are trying to log information from project startup, it would be
reasonable to make the logs INFO or higher, as DEBUG level logs could be
excluded even with a user setting of verbose.