Update Android Studio version (#8565)
I removed code that seems to be related to making gradle files editable
within a Flutter project window as the Android Studio code has changed.
I will record some additional details in my internal notes.

There were also test compilation failures that suggested the annotations
library wasn't required correctly. (see
https://stackoverflow.com/questions/62066166/string-notnull-syntax-causes-notnull-not-applicable-to-type-use-in-in
- thanks @pq). When I asked gemini about this, it suggested that we were
compiling production source code twice for tests, because the test
source sets included `src` as well. I guess it turns out that this isn't
needed and the dependency on the production source code is implied;
additionally, compiling the production source during test is bad because
it can be compiled in the wrong environment. I think this is
documentation that includes this info if we want to investigate further
at a later time:
https://docs.gradle.org/current/userguide/building_java_projects.html#sec:java_source_sets