Add config for gemini code assist in the repository (#8607)
This follows
https://developers.google.com/gemini-code-assist/docs/customize-gemini-behavior-github
for adding some settings before we enable code assist here.

The style guide text is copied from the documentation, as I don't think
we have any codified Java conventions at the moment. We can add more to
this file in the future though, as it is meant to be human-readable
instructions.